extracted some utiltiy functions
